refactor($ALL): 폴더 이름 변경
폴더 이름의 앞문자를 소문자로 변경, 패키지 변경에 따른 리팩터링
