+++ src/main/java/calculator/Application.java
@@ -1,6 +1,6 @@
 package calculator;
 
-import calculator.Controller.CalculatorController;
+import calculator.controller.CalculatorController;
 
 public class Application {
     public static void main(String[] args) {
