+++ src/main/java/calculator/constants/InputMessages.java
@@ -1,4 +1,4 @@
-package calculator.Constants;
+package calculator.constants;
 
 public class InputMessages {
     public final static String INSERT_STRING = "덧셈할 문자열을 입력해주세요";
