+++ src/test/java/calculator/model/DelimiterTest.java
@@ -1,4 +1,4 @@
-package calculator.Model;
+package calculator.model;
 
 import calculator.Application;
 import camp.nextstep.edu.missionutils.test.NsTest;
@@ -7,12 +7,13 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-public class NumberTest extends NsTest {
+public class DelimiterTest extends NsTest {
+
     @ParameterizedTest
-    @DisplayName("불가능한 숫자 객체를 확인합니다.")
-    @ValueSource(strings = {".", ";", ":", "//", "\\\\n", "-1.0", "-1", "0", "-1020394"})
-    void number_테스트(String input) {
-        Assertions.assertThatThrownBy(() -> new Number(input))
+    @DisplayName("불가능한 구분자를 확인합니다.")
+    @ValueSource(strings = {"6", "", " ", ".", "19.0"})
+    void delimiter_테스트_false(String input) {
+        Assertions.assertThatThrownBy(() -> new Delimiter(input))
                 .isInstanceOf(IllegalArgumentException.class);
     }
