+++ src/test/java/calculator/utils/NumberUtilsTest.java
@@ -1,4 +1,4 @@
-package calculator.Utils;
+package calculator.utils;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -7,7 +7,6 @@ import camp.nextstep.edu.missionutils.test.NsTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-
 
 public class NumberUtilsTest extends NsTest {
